requirements.txt
typer
requests
youtube-search-python
jq
pyyaml